+++ policy_sentry/command/write_policy.py
@@ -6,13 +6,13 @@ import json
 import logging
 import yaml
 import click
-from policy_sentry.util.logging import set_log_level
 from policy_sentry.shared.constants import DATABASE_FILE_PATH
 from policy_sentry.shared.database import connect_db
 from policy_sentry.util.file import read_yaml_file
 from policy_sentry.writing.sid_group import SidGroup
-
+import click_log
 logger = logging.getLogger()
+click_log.basic_config(logger)
 
 
 @click.command(
@@ -32,18 +32,12 @@ logger = logging.getLogger()
     help="Minimize the resulting statement with *safe* usage of wildcards to reduce policy length. "
     "Set this to the character length you want - for example, 4",
 )
-@click.option(
-    "--log-level",
-    help="Set the logging level. Choices are CRITICAL, ERROR, WARNING, INFO, or DEBUG. Defaults to INFO.",
-    type=click.Choice(["CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG"]),
-    default="INFO",
-)
-def write_policy(input_file, minimize, log_level):
+@click_log.simple_verbosity_option(logger)
+def write_policy(input_file, minimize):
     """
     Write a least-privilege IAM Policy by supplying either a list of actions or
     access levels specific to resource ARNs!
     """
-    set_log_level(logger, log_level)
 
     db_session = connect_db(DATABASE_FILE_PATH)