+++ policy_sentry/command/write_policy_dir.py
@@ -13,10 +13,10 @@ from policy_sentry.util.file import (
     write_json_file,
     check_valid_file_path,
 )
-from policy_sentry.util.logging import set_log_level
 from policy_sentry.shared.constants import DATABASE_FILE_PATH
-
+import click_log
 logger = logging.getLogger()
+click_log.basic_config(logger)
 
 
 @click.command(
@@ -41,18 +41,11 @@ logger = logging.getLogger()
     type=int,
     help="Minimize the resulting statement with *safe* usage of wildcards to reduce policy length. Set this to the character length you want - for example, 4",
 )
-@click.option(
-    "--log-level",
-    help="Set the logging level. Choices are CRITICAL, ERROR, WARNING, INFO, or DEBUG. Defaults to INFO.",
-    type=click.Choice(["CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG"]),
-    default="INFO",
-)
-def write_policy_dir(input_dir, output_dir, minimize, log_level):
+@click_log.simple_verbosity_option(logger)
+def write_policy_dir(input_dir, output_dir, minimize):
     """
     write_policy, but this time with an input directory of YML/YAML files, and an output directory for all the JSON files
     """
-    set_log_level(logger, log_level)
-
     db_session = connect_db(DATABASE_FILE_PATH)
     input_dir = os.path.abspath(input_dir)
     output_dir = os.path.abspath(output_dir)